Changed _Date type to long
long type is more suited to represent dates: e.g. the excel format for date is long

+++ src/Derivative.h
@@ -3,7 +3,7 @@
 
 #include "Underlying.h"
 
-using _Date = int;
+using _Date = long;
 
 /**
  * TODO:
